==> 3. predict.ipynb <==
import os
import numpy as np
import pandas as pd
from glob import glob

import joblib

from utils import get_features, text_embedding

from keras.models import load_model
from sentence_transformers import SentenceTransformer

current_path=os.getcwd().replace("\\", "/").replace("c:", "C:")
data_path=current_path+'/data/'
model_path=current_path+"/model/"

model = load_model(model_path+'my_model.h5')
encoder = joblib.load(model_path+'encoder.pkl')
scaler = joblib.load(model_path+'scaler.pkl')
with open(model_path+'text_model_name.txt', 'r') as file:
    text_model_name = file.read().strip()

txt_data=pd.read_csv(data_path+"label/original.csv")

txt_data[txt_data["final_label"]=="happiness"].sample(3)

file_name='5f61ec172e23c7161accd3ef'
new_audio_path = data_path+"/wavFiles/"+file_name+'.wav'
new_audio_features = get_features(new_audio_path)
new_audio_features_df = pd.DataFrame([new_audio_features])

new_audio_features_df['wav_id']=file_name
new_audio_features_df['final_label']=txt_data[txt_data['wav_id']==file_name]['final_label'].values[0].lower()
new_audio_features_df['sentence']=txt_data[txt_data['wav_id']==file_name]['sentence'].values[0]

txt_embed = text_embedding(model_name = text_model_name)
new_audio_features_df = txt_embed.transform(new_audio_features_df)

# 특징 표준화
new_audio_features_scaled = scaler.transform(new_audio_features_df)

# 모델에 맞게 차원 확장
new_audio_features_scaled = np.expand_dims(new_audio_features_scaled, axis=2)

# 예측 수행
prediction = model.predict(new_audio_features_scaled)
predicted_label = np.argmax(prediction, axis=1)

predicted_label

num_categories = encoder.categories_[0].size  # OneHotEncoder에 의해 인코딩된 카테고리 수

# 모델의 예측 결과를 원-핫 인코딩 형식으로 변환
predicted_one_hot = np.zeros((predicted_label.size, num_categories))
predicted_one_hot[np.arange(predicted_label.size), predicted_label] = 1

# 원-핫 인코딩된 예측 결과를 실제 레이블로 변환
actual_label = encoder.inverse_transform(predicted_one_hot)


actual_label[0][0]

# 출력 형식 설정
np.set_printoptions(precision=6, suppress=True)

prediction*100

prediction
