==> 4. api.ipynb <==
import pandas as pd
import os

current_path=os.getcwd().replace("\\", "/").replace("c:", "C:")
data_path=current_path+'/data/'

label_file=data_path+'label/original.csv'
label=pd.read_csv(label_file)

label[label["wav_id"]=='5e2ac3d55807b852d9e01fd6']

import requests

# API 엔드포인트
url = 'http://127.0.0.1:8000/predict/'

# 보낼 데이터
data = {
    'audioFilePath': "C:/Users/82105/Desktop/코드/korean_emotions/data/wavFiles/5e2ac3d55807b852d9e01fd6.wav",  # 오디오 파일 경로
    'sentence': "우리 아빠 어제 술먹고 또 사고쳤어"  # 문장
}

# POST 요청 보내기
response = requests.post(url, json=data)

# 응답 출력
if response.status_code == 200:
    print("Response from server:", response.json())
else:
    print("Error:", response.status_code, response.text)
